+++ bosesoundtouchapi/soundtouchmodelrequest.py
@@ -22,10 +22,15 @@ class SoundTouchModelRequest:
     @abstractmethod
     def ToXmlRequestBody(self, encoding:str='utf-8') -> str:
         """ 
-        Returns a POST request body for updating a configuration.
+        Returns a POST request body, which is used to update the device configuration.
         
+        Args:
+            encoding (str):
+                encode type (e.g. 'utf-8', 'unicode', etc).  
+                Default is 'utf-8'.
+
         Returns:
-            An xml string that can be used in a POST request to update a
+            An xml string that can be used in a POST request to update the
             device configuration.
         """
         raise NotImplementedError("The '%s' class has not overridden the RequestBody() method." % (self.__name__))